==> notebooks/01_data_quality_checks.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import datetime as datetime
import warnings
warnings.filterwarnings('ignore')
import os

df = pd.read_csv("./data/raw/policy_data.csv")
df.shape  # to check the shape of the dataframe

# Display the first few rows of the dataframe
df.head()

# Display summary information about the dataframe and daata types
df.info()

df.describe()

df.dtypes

# Check for missing values in each column
df.isnull().sum()

# check policy _id 
df['policy_id'].is_unique
df.columns

#exposure are negative and more that 1 and it is not hence drop exposure
df = df.drop(columns=['exposure'],axis=1)
df.head()

df['channel'].value_counts()

# Check for negative values in 'sum_insured' column
(df['sum_insured'] < 0).sum()

df_clean = df[df['sum_insured'] > 0].copy()
df_clean.shape

(df_clean['sum_insured'] < 0).sum()
df_clean.dtypes

# Parse dates first
df_clean['policy_start_date'] = pd.to_datetime(df['policy_start_date'])
df_clean['policy_end_date'] = pd.to_datetime(df['policy_end_date'])

# check valid dates and remove invalid dates
df_clean_valid = df_clean[df_clean['policy_start_date'] < df_clean['policy_end_date']]
df_clean.shape

# Calculate exposure in years and earned premium
df_clean_valid['exposure'] = ((df_clean_valid['policy_end_date'] - df_clean_valid['policy_start_date']).dt.days + 1) / 365.25
df_clean_valid['exposure'] = df_clean_valid['exposure'].clip(lower=0, upper=1)
df.shape
df_clean_valid['earned_premium'] = df_clean_valid['premium'] * df_clean_valid['exposure']

# Save cleaned dataframe
df_clean_valid.to_csv('./data/processed/policy_clean.csv', index=False)
print("✅ Saved: policy_clean")

#clean claims data
# Load claims data
df = pd.read_csv('./data/raw/claims_data.csv')
df.shape

df.info()

df = df.iloc[:,[0,1,2,3,4,5]]
df.info()

# Check for missing values
(df['claim_amount'] <= 0).sum()

df.head()

df['policy_id'].is_unique

df['accident_date'] = pd.to_datetime(df['accident_date'], dayfirst=True , errors='coerce')
df.dtypes

df = df.astype({
    'claim_id': 'Int64',
    'accident_year': 'Int64'
})


df.dtypes

# Save cleaned dataframe
df.to_csv('./data/processed/claims_clean.csv', index=False)
print("✅ Saved: claims_clean")

#cleaning and loading payment_data
payment = pd.read_csv('./data/raw/payment_data.csv').copy()

payment.info()

payment['payment_date'] = pd.to_datetime(payment['payment_date'], dayfirst=True , errors='coerce')
payment.dtypes

#checking Negative values
print((payment['case_reserve'] < 0).sum())
print((payment['paid_amount'] < 0).sum())

# Save cleaned dataframe
payment.to_csv('./data/processed/payment_clean.csv', index=False)
print("✅ Saved: payment_clean")
