==> notebooks/02_exposure_calculation.ipynb <==
import numpy as np
import pandas as pd
import os
from datetime import datetime
os.getcwd()

df = pd.read_csv('policy_clean.csv')


df.head()

#check summary statistics for 'exposure' column and look for anomalies
df['exposure'].describe()

#removing exposure and calculating it again from policy start and end date
df = df.drop(columns=['exposure'])
df.info()

#convert policy_start_date and policy_end_date to datetime format
df['policy_start_date'] = pd.to_datetime(df['policy_start_date'], dayfirst=True , errors='coerce')
df['policy_end_date'] = pd.to_datetime(df['policy_end_date'], dayfirst=True , errors='coerce')
df.dtypes

# Calculate exposure in years
df['exposure'] = ((df['policy_end_date'] - df['policy_start_date']).dt.days + 1) / 365
df['earned_premium'] = df['premium'] * df['exposure']


# Extract policy year
df['policy_year'] = (df['policy_start_date'].dt.year).astype('int16')
df['exposure'] = df['exposure'].clip(lower=0, upper=1)
df.shape



df['exposure'].describe()

df.to_csv('exposure_table.csv', index=False)
