==> notebooks/03_claims_linkage.ipynb <==
import pandas as pd
import numpy as np
import datetime as datetime

policy_df = pd.read_csv('./data/processed/policy_clean.csv',usecols = ['policy_id','policy_start_date','policy_end_date','vehicle_age','region','channel','sum_insured','premium','exposure'])
claims_df= pd.read_csv('./data/processed/claims_clean.csv',usecols = ['policy_id','claim_id','accident_date','claim_amount','development_year'])
payment_df = pd.read_csv('./data/processed/payment_clean.csv')


policy_df.dtypes


claims_df.dtypes

# Convert date columns to datetime format
policy_df['policy_start_date'] = pd.to_datetime(policy_df['policy_start_date'], errors='coerce')
policy_df['policy_end_date'] = pd.to_datetime(policy_df['policy_end_date'],  errors='coerce')
claims_df['accident_date'] = pd.to_datetime(claims_df['accident_date'], errors='coerce')

# merge policy And claims data on policy number
merged_df = pd.merge(policy_df,claims_df, on='policy_id', how='left')
merged_df.columns

merged_df.info()

#valid claims within policy period and invalid claims outside policy period
valid_claims = merged_df[
    (merged_df['accident_date'] >= merged_df['policy_start_date']) &
    (merged_df['accident_date'] <= merged_df['policy_end_date'])]

invalid_claims = merged_df[
    (merged_df['accident_date'] < merged_df['policy_start_date']) |
    (merged_df['accident_date'] > merged_df['policy_end_date'])]


# Link valid claims with exposure table
linked_valid_claims=valid_claims.copy()
linked_valid_claims

# Calculate claim frequency
claim_counts = linked_valid_claims.groupby('policy_id').size().reset_index(name='claim_count')
frequency_df = pd.merge(linked_valid_claims, claim_counts, on='policy_id', how='left')

claim_counts

frequency_df

frequency_df['policy_year'] = frequency_df['policy_start_date'].dt.year
frequency_df.columns

frequency_dfs = frequency_df[['policy_id','claim_id' ,'exposure', 'claim_count', 'vehicle_age', 'region', 'channel', 'sum_insured','policy_year']].copy()

#frequency data and severity data export for GLM Inputs
frequency_dfs.to_csv('./data/processed/frequency_data.csv', index=False)
severity_df = linked_valid_claims[['policy_id','claim_id','accident_date', 'claim_amount', 'vehicle_age', 'region', 'channel', 'sum_insured']]
severity_df.to_csv('./data/processed/severity_data.csv', index=False)

#reserving triangle
rt_df = pd.merge(frequency_df,payment_df[['claim_id','paid_amount','case_reserve']], on='claim_id', how='left')
rt_df.columns

rt_df['accident_year']= rt_df['accident_date'].dt.year
rt_df.columns

reserving_df = rt_df[['accident_year','development_year','paid_amount','case_reserve']]


reserving_df.to_csv('./data/processed/reserving_triangle_data.csv', index=False )
