==> notebooks/04_frequency_glm.ipynb <==
import pandas as pd
import numpy as np
import statsmodels.api as sm
import statsmodels.formula.api as smf
import os

fdf = pd.read_csv('./data/processed/frequency_data.csv')

fdf.columns

 # Fit a Poisson GLM for claim frequency with exposure as an offset 
model = smf.glm( formula="claim_count ~ vehicle_age + region + channel + sum_insured",
                 data=fdf, family=sm.families.Poisson(), offset=np.log(fdf["exposure"]) ).fit()
print(model.summary())


# Predict claim frequency
fdf['predicted_frequency'] = model.predict(fdf)

fdf.head()

fdf.to_csv('./data/processed/frequency_output.csv', index=False)

import matplotlib.pyplot as plt

coef = model.params.drop("Intercept", errors="ignore")
relativity = np.exp(coef)

plt.figure(figsize=(10,6))
relativity.sort_values().plot(kind="barh")
plt.axvline(1, linestyle="--")
plt.title("GLM Risk Relativities")
plt.xlabel("Relativity")
plt.ylabel("Rating Factors")
plt.tight_layout()
plt.savefig("./visuals/fre_relativity_plots.png", dpi=300)
plt.close()
