==> notebooks/05_severity_glm.ipynb <==
import pandas as pd
import numpy as np
import statsmodels.api as sm
import statsmodels.formula.api as smf
import os
os.getcwd()
os.chdir('d:/datasets/d1/Actuarial-Pricing-and-Reserving-GLM/')

sdf = pd.read_csv('./data/processed/severity_data.csv')

sdf.columns

# Fit a Gamma GLM for claim severity
sdf_model = smf.glm(
    formula="claim_amount ~ vehicle_age + region + channel + sum_insured",
    data=sdf,
    family=sm.families.Gamma(sm.families.links.Log())
).fit()

print(sdf_model.summary())

# Predict claim severity
sdf['predicted_severity'] = sdf_model.predict()

sdf.head()

sdf.to_csv('./data/processed/severity_output.csv', index=False)
