==> notebooks/06_pricing_output.ipynb <==
import pandas as pd
import numpy as np
import statsmodels.api as sm
import statsmodels.formula.api as smf
import os

policy = pd.read_csv('./data/processed/policy_clean.csv')

policy.columns

# Merge frequency and severity predictions into pricing table
freq_pred = pd.read_csv('./data/processed/frequency_output.csv')
freq_pred.columns




sev_pred = pd.read_csv('./data/processed/severity_output.csv') 
sev_pred.columns


policy = policy.merge(freq_pred[['policy_id', 'predicted_frequency']], on='policy_id')
policy = policy.merge(sev_pred[['policy_id', 'predicted_severity']], on='policy_id')  

policy.columns

# Select relevant columns for final pricing output
pricing = policy[['policy_id','vehicle_age','region','channel','sum_insured','predicted_frequency','predicted_severity']]

#loading addesd columns and calculating final premium
# 1.40 = expense + profit + risk margin(.25+.10+.05)
pricing["pure_premium"] = pricing["predicted_frequency"] * pricing["predicted_severity"]
expense = 0.25
profit = 0.10
risk_margin = 0.05

loading_factor = 1 + expense + profit + risk_margin
pricing["final_premium"] = pricing["pure_premium"] * loading_factor
pricing.columns


pricing.to_csv('./data/processed/pricing_output.csv', index=False)
